REFACTORIZACION EN LOS SERVICIOS

+++ colegio/src/main/java/com/api/colegio/repository/EstudianteRepository.java
@@ -3,7 +3,7 @@ package com.api.colegio.repository;
 import com.api.colegio.model.Estudiante;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstudianteRepository extends JpaRepository<Estudiante, Long> {
+public interface EstudianteRepository extends JpaRepository<Estudiante, Integer> {
